+++ src/app/locales/en/utils/selector-filesystem-browser-model.js
@@ -1,0 +1,8 @@
+import FilesystemBrowserModel from './filesystem-browser-model';
+import _ from 'lodash';
+
+export default _.merge({}, FilesystemBrowserModel, {
+  fileActions: {
+    chooseCurrentDir: 'Choose current directory',
+  },
+});